requirements.txt
pip>=24.3.1
setuptools>=68.0.0
wheel>=0.40.0
streamlit>=1.25
pandas==1.5.3
numpy==1.24.2
torch==2.0.1
torchvision==0.15.2
tqdm
transformers==4.32.0
pillow
matplotlib
langchain==0.3.14
langchain-community==0.3.14
langchain-ollama==0.2.2
langcodes==3.3.0
openai==0.28.0
chromadb==0.4.18 
pysqlite3-binary 
tiktoken==0.5.2